backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
django-ratelimit==4.1.0
mongoengine==0.29.1
pymongo==4.10.1
channels==4.2.0
channels-redis==4.2.1
celery==5.4.0
redis==5.2.1
pandas==2.2.3
openpyxl==3.1.5
Pillow==11.1.0
python-dotenv==1.0.1
django-redis==5.4.0
gunicorn==23.0.0
whitenoise==6.9.0
reportlab==4.3.1
Werkzeug==3.1.3
